refactor : separate ProfileDropdown Component from Header
